groups bruker outline av radio og check
